Use futurize to have python 3 compatible int instead of long

+++ usr/lib/enigma2/python/Screens/InfoBarGenerics.py
@@ -2412,7 +2412,7 @@ class InfoBarCueSheetSupport:
         r = seek.getPlayPosition()
         if r[0]:
             return None
-        return long(r[1])
+        return int(r[1])
 
 
     def cueGetEndCutPosition(self):
